# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/posts": {
      "get": {
        "operationId": "PostController_getAllPosts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "전체 게시물 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Many"
                  }
                }
              }
            }
          }
        },
        "summary": "전체 게시물 조회",
        "tags": [
          "Post"
        ]
      },
      "post": {
        "operationId": "PostController_createPost",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Create"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "게시물 작성 완료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Simple"
                }
              }
            }
          }
        },
        "summary": "게시물 작성",
        "tags": [
          "Post"
        ]
      }
    },
    "/posts/{id}/only-post": {
      "get": {
        "operationId": "PostController_getSinglePost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "게시물 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "단일 게시물 데이터",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Single"
                }
              }
            }
          }
        },
        "summary": "단일 게시물 조회",
        "tags": [
          "Post"
        ]
      }
    },
...
    "/posts/{id}/like": {
      "post": {
        "operationId": "PostController_postLike",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "게시물 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "좋아요 반영 결과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Like"
                }
              }
            }
          }
        },
        "summary": "게시물 좋아요/취소 토글",
        "tags": [
          "Post"
        ]
      }
    },
    "/posts/{postId}": {
      "patch": {
        "operationId": "PostController_updatePost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "게시물 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "게시물 수정 완료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Update"
                }
              }
            }
          }
        },
        "summary": "게시물 수정",
        "tags": [
          "Post"
        ]
      },
      "delete": {
        "operationId": "PostController_deletePost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "게시물 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "게시물 및 관련 댓글/좋아요 삭제 완료"
          }
        },
        "summary": "게시물 삭제 (관련 댓글, 좋아요 포함)",
        "tags": [
          "Post"
        ]
      }
    },
    "/comments/{id}": {
      "get": {
        "operationId": "CommentController_getComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Comment"
        ]
      }
    },
...
    "/comments/reply": {
      "get": {
        "operationId": "CommentController_getReplyCommentsByParent",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Comment"
        ]
      },
      "post": {
        "operationId": "CommentController_createReplyComment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReply"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Comment"
        ]
      }
    },
...
    "/comments": {
      "post": {
        "operationId": "CommentController_createComment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Create"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Comment"
        ]
      },
      "patch": {
        "operationId": "CommentController_updateComment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Comment"
        ]
      },
      "delete": {
        "operationId": "CommentController_deleteComment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Comment"
        ]
      }
    },
    "/comments/like": {
      "post": {
        "operationId": "CommentController_commentLike",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Comment"
        ]
      }
    },
...
    "/record/{userId}/admin": {
      "get": {
        "operationId": "RecordController_getUserRecords",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Record"
        ]
      }
    },
...
  },
  "info": {
    "title": "루파봇 API",
    "description": "루파봇 커뮤니티 + 상점 + 대전 API 문서",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Many": {
        "type": "object",
        "properties": {}
      },
      "Single": {
        "type": "object",
        "properties": {}
      },
...
      "Create": {
        "type": "object",
        "properties": {}
      },
      "Simple": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "string",
            "example": "660fa04149ac491601f005b6"
          }
        },
        "required": [
          "postId"
        ]
      },
      "Like": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "liked"
          }
        },
        "required": [
          "status"
        ]
      },
      "Update": {
        "type": "object",
        "properties": {}
      },
      "CreateReply": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}